--- a/dozer-api/src/rest/tests/auth.rs
+++ b/dozer-api/src/rest/tests/auth.rs
@@ -70,16 +70,7 @@ async fn check_status(
         security,
         CorsOptions::Permissive,
         vec![Arc::new(
-            CacheEndpoint::new(
-                &*cache_manager,
-                test_utils::get_schema().0,
-                endpoint.clone(),
-                test_utils::get_log_path().as_path(),
-                None,
-                None,
-            )
-            .unwrap()
-            .0,
+            CacheEndpoint::open(&*cache_manager, endpoint.clone()).unwrap(),
         )],
     );
     let app = actix_web::test::init_service(api_server).await;
@@ -101,23 +92,13 @@ async fn _call_auth_token_api(
     body: Option<Value>,
 ) -> ServiceResponse<impl MessageBody> {
     let endpoint = test_utils::get_endpoint();
-    let schema = test_utils::get_schema();
     let schema_name = endpoint.name.clone();
     let cache_manager = test_utils::initialize_cache(&schema_name, None);
     let api_server = ApiServer::create_app_entry(
         Some(ApiSecurity::Jwt(secret)),
         CorsOptions::Permissive,
         vec![Arc::new(
-            CacheEndpoint::new(
-                &*cache_manager,
-                schema.0,
-                endpoint,
-                test_utils::get_log_path().as_path(),
-                None,
-                None,
-            )
-            .unwrap()
-            .0,
+            CacheEndpoint::open(&*cache_manager, endpoint).unwrap(),
         )],
     );
     let app = actix_web::test::init_service(api_server).await;