--- a/dozer-api/src/rest/tests/routes.rs
+++ b/dozer-api/src/rest/tests/routes.rs
@@ -30,16 +30,7 @@ async fn list_route() {
         None,
         CorsOptions::Permissive,
         vec![Arc::new(
-            CacheEndpoint::new(
-                &*cache_manager,
-                test_utils::get_schema().0,
-                endpoint.clone(),
-                test_utils::get_log_path().as_path(),
-                None,
-                None,
-            )
-            .unwrap()
-            .0,
+            CacheEndpoint::open(&*cache_manager, endpoint.clone()).unwrap(),
         )],
     );
     let app = actix_web::test::init_service(api_server).await;
@@ -98,16 +89,7 @@ async fn count_and_query_route() {
         None,
         CorsOptions::Permissive,
         vec![Arc::new(
-            CacheEndpoint::new(
-                &*cache_manager,
-                test_utils::get_schema().0,
-                endpoint.clone(),
-                test_utils::get_log_path().as_path(),
-                None,
-                None,
-            )
-            .unwrap()
-            .0,
+            CacheEndpoint::open(&*cache_manager, endpoint.clone()).unwrap(),
         )],
     );
     let app = actix_web::test::init_service(api_server).await;
@@ -152,16 +134,7 @@ async fn get_route() {
         None,
         CorsOptions::Permissive,
         vec![Arc::new(
-            CacheEndpoint::new(
-                &*cache_manager,
-                test_utils::get_schema().0,
-                endpoint.clone(),
-                test_utils::get_log_path().as_path(),
-                None,
-                None,
-            )
-            .unwrap()
-            .0,
+            CacheEndpoint::open(&*cache_manager, endpoint.clone()).unwrap(),
         )],
     );
     let app = actix_web::test::init_service(api_server).await;