Otherwise one of the tokio worker threads is blocked and can't be used for other tasks.

Also, tokio runtime will block forever when shutdown because we never yield.