@chubei 2023-04-11
# `main`

```text
cargo run --release --example reader  0.43s user 0.12s system 66% cpu 0.834 total
```

# PR

```text
cargo run --release --example reader  0.43s user 0.16s system 66% cpu 0.886 total
```

Minor performance degrade.